fix: reference to TeletextConfig
