+++ src/rust/lib_ccxr/src/common/mod.rs
@@ -18,8 +18,6 @@
 
 mod constants;
 mod options;
-mod teletext;
 
 pub use constants::*;
 pub use options::*;
-pub use teletext::*;
