+++ src/rust/src/common.rs
@@ -14,9 +14,9 @@ use lib_ccxr::common::OutputFormat;
 use lib_ccxr::common::SelectCodec;
 use lib_ccxr::common::StreamMode;
 use lib_ccxr::common::StreamType;
-use lib_ccxr::common::TeletextConfig;
 use lib_ccxr::hardsubx::ColorHue;
 use lib_ccxr::hardsubx::OcrMode;
+use lib_ccxr::teletext::TeletextConfig;
 use lib_ccxr::time::units::Timestamp;
 use lib_ccxr::time::units::TimestampFormat;
 use lib_ccxr::util::encoding::Encoding;
